Be able to build the logger on windows
Seems like provider can build for windows for the kairosctl command and
it would fail due to the systemd libs not being available to build under
windows so we need to protect against that
