+++ types/logger_windows.go
@@ -1,0 +1,13 @@
+//go:build windows
+
+package types
+
+import "io"
+
+func isJournaldAvailable() bool {
+	return false
+}
+
+func getJournaldWriter() io.Writer {
+	return nil // No journald on Windows
+}
